% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.data.frame.mzR}
\alias{as.data.frame.mzR}
\title{Kopnvertiere ein \code{mzR} Objekt in ein Tabelle}
\usage{
\method{as.data.frame}{mzR}(x, ...)
}
\arguments{
\item{x}{Ein Objekt der Klasse \code{mzR}.}

\item{...}{Ungenuetzt}
}
\description{
Generische Funktion \link{as.data.frame} implementiert fuer \code{mzR} Objekte.
}
