% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{Design}
\alias{Design}
\title{Hilfsfunktion zur Verwendung des survey-Pakets.}
\usage{
Design(x)
}
\arguments{
\item{x}{MZ Daten - Output von Funktion ImportData()}
}
\value{
Under construction.
}
\description{
Erstellt ein svyrepdesign Objekt aus den eingelesen Daten
}
\details{
Under construction.
}
\examples{

\dontrun{
dat <- ImportData(year=2014,quarter=4, comp_diff=TRUE, comp_diff_lag=4)
des <- Design(dat)
}

}
\seealso{
\code{\link{ImportData},\link{IndivImportData}}
}

