% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFehler.R
\name{export}
\alias{export}
\alias{export.mzR}
\title{Export der Schaetz-und Fehlerrechnungsergebnisse.}
\usage{
export(x, outFilePath = getwd(), outFileName = NULL)

\method{export}{mzR}(x, outFilePath = getwd(), outFileName = NULL)
}
\arguments{
\item{x}{Ergebnis von GroupSize, GroupRate, Total oder Mean.}

\item{outFilePath}{Pfad, wo die CSV-Datei gespeichert werden soll. Default
ist das aktuelle Work Directory.}

\item{outFileName}{Name der CSV-Datei in die exportiert werden soll.}
}
\value{
Ein .csv-File mit den Ergebnissen wird in \code{out_file_pfad}
abgespeichert.
}
\description{
Funktion exportiert die Ergebnisse der Schaetz-und Fehlerrechnungsfunktionen
in ein .csv-File.
}
\examples{
# Daten laden (oder Daten einlesen mit ImportData() bzw. IndivImportData())
data(mzTestData)

# Arbeitslosenzahlen (Absolutwerte)
ergebnis <- GroupSize(mzTestData,TFstring="xerwstat==2&balt>=15&balt<=74")
\dontrun{
# Arbeitslosenzahlen (Absolutwerte) ins Working Directory (zu erfragen mit getwd()) exportieren
export(ergebnis)
# Erwerbsstatus nach Bundesland ins Working Directory exportieren
export(GroupSize(dat,each="xerwstat+xnuts2"))
}

}
\seealso{
\code{\link{ImportData},\link{IndivImportData},\link{GetLabels},\link{GroupSize},\link{GroupRate},\link{Total},\link{Mean}}
}
