% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLabels.R
\name{GetLabels}
\alias{GetLabels}
\title{Labels einer Variablen wiedergeben.}
\usage{
GetLabels(x, var, curr = TRUE)
}
\arguments{
\item{x}{MZ Daten - Output von Funktion \link{ImportData} oder \link{IndivImportData}.}

\item{var}{Character Vektor mit Variable(n) fuer die die Labels angezeigt
werden sollen.}

\item{curr}{TRUE/FALSE ob \code{var} auf "current" oder "previous" Zeitpunkt
bezogen sein soll, d.h. falls \code{comp_diff_lag} ungleich NULL in der
Funktion \code{ImportData}, koennen mit \code{curr=FALSE} auch die Labels
des weniger aktuellen der beiden Datensaetze ausgegeben werden.
\code{curr=TRUE} bezieht sich auf den aktuelleren der beiden Zeitpunkte.}
}
\value{
Output sind die Labels einer Variablen \code{var} falls im
dg7-Datensatz welche definiert wurden.
}
\description{
Funktion zeigt SPSS-Labels einer Variablen an.
}
\examples{
# Daten laden (oder Daten einlesen mit ImportData() bzw. IndivImportData())
data(mzTestData)

# Labels von Variable xerwstat abfragen
GetLabels(mzTestData,"xerwstat")
# Labels von Variablen dtstd und xnuts2 abfragen
GetLabels(mzTestData,c("dtstd","xnuts2"))

}
\seealso{
\code{\link{ImportData},\link{IndivImportData},\link{ImportAndMerge}}
}

