% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{getReplicates}
\alias{getReplicates}
\title{Extrahiere die Bootstrap Replikla aus einem mzR Objekt}
\usage{
getReplicates(x)
}
\arguments{
\item{x}{Ein Objekt der Klasse \code{mzR}. Typischerweise generiert durch \code{GroupRate}, \code{Groupsize},
\code{Median}, \code{Mean} oder \code{Total}.}
}
\value{
Tabelle mit Bootstrapreplika. Die Spalten entsprechen den Gruppierungsvariablen der
Auswertung, falls vorhanden (siehe das Argument \code{each} in \code{GroupRate}, \code{Median}, etc.). Die Zeilen
enstsprechen den Bootstrapgewichten (typischerweise 500 Stück).
}
\description{
Falls ein \code{mzR} Objekt mit dem Parameter \code{replicate = TRUE} erzeugt wurde, lassen sich mit dieser
Funktion alle Schätzwerte zurückgeben.
}
