% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSize.R
\name{GroupRate}
\alias{GroupRate}
\alias{GroupSize}
\title{Schaetz- und Fehlerrechnungsfunktionen.}
\usage{
GroupRate(
  x,
  TFstring,
  TFstring2 = NULL,
  each = NULL,
  byeach = TRUE,
  thousands_separator = TRUE,
  digits = 2,
  replicates = FALSE,
  add_sample_size = FALSE
)

GroupSize(
  x,
  TFstring = NULL,
  each = NULL,
  thousands_separator = TRUE,
  digits = 2,
  replicates = FALSE,
  add_sample_size = FALSE
)
}
\arguments{
\item{x}{MZ Daten - Output von Funktion \link{ImportData}.}

\item{TFstring}{Character oder NULL: Logische Einschraenkung der Gruppe, im Fall der Anteile (\code{GroupRate}):
Einschraenkung fuer Zaehler. Falls NULL, gilt keine Einschraenkung.}

\item{TFstring2}{Character oder NULL: Logische Einschraenkung fuer den Nenner.
Falls NULL, gilt keine Einschraenkung fuer den Nenner ausser wenn \code{each} ungleich NULL und \code{byeach=TRUE}.}

\item{each}{Character oder NULL: Name der Variable nach der getrennt berechnet werden soll.
Hier koennen auch mehrere Variablen angegeben werden und zwar in einem character string, getrennt durch ein +, siehe Examples.}

\item{byeach}{Logical: Vorgabe fuer den Nenner. \code{byeach} bezieht sich auf den
Parameter \code{each} und gibt an, ob bei \code{GroupRate} zusaetzlich zu
\code{TFstring2} auch die jeweilige durch \code{each} definierte Einschraenkung
in den Nenner kommen soll, also ob die Raten jeweils bezogen auf die Auspraegungen
bzw. Auspraegungskombinationen von \code{each} berechnet werden sollen (\code{byeach=TRUE})
oder rein bezogen auf \code{TFstring2} (\code{byeach=FALSE}).}

\item{thousands_separator}{Logical: Wenn TRUE, werden Tausendertrennzeichen
angezeigt.}

\item{digits}{Numerischer Wert: Anzahl der Nachkommastellen im angezeigten Ergebnis. Default
ist 2.}

\item{replicates}{Fürge einen Vektor aus Schätzwerten zum Output hinzu? Die Anzahl der Schätzwerte
pro Gruppe in \code{each} entspricht der Anzahl der Bootstrapreplikate (typischerweise 500).
Siehe auch \link{getReplicates}.}

\item{add_sample_size}{Logical: ob Stichprobengroesse der jeweiligen Gruppe auch ausgegeben werden soll.}
}
\value{
Output ist ein Objekt der Klasse \code{mzR}.
}
\description{
Funktion berechnet absolute (\code{GroupSize}) oder relative
(\code{GroupRate}) Schaetzwerte und die zugehoerigen Fehler.
}
\details{
Wiedergegeben wird der Schaetzer \code{est}, der Stichprobenfehler
\code{sd}, der Variationskoeffizient \code{cv} und die untere/obere Grenze
des 95\\% Konfidenzintervalls \code{cil_2.5\%}/\code{ciu_97.5\%}.
Die Fehler werden mit Hilfe von Replikationsgewichten aus einem Bootstrapverfahren
berechnet, d.h. \code{sd} entspricht der Standardabweichung der mit den Bootstrapgewichten
berechneten Schaetzwerte, \code{cil_2.5\%} und \code{ciu_97.5\%} sind die
entsprechenden 2.5\\% und 97.5\\% Quantile und \code{cv=sd/est}.
}
\examples{
# Daten laden (oder Daten einlesen mit ImportData() bzw. IndivImportData())
data(mzTestData)

# Arbeitslosenzahlen: Absolutwerte und Veraenderung
GroupSize(mzTestData,TFstring="xerwstat==2&balt>=15&balt<=74")
# Arbeitslosenquoten: Prozentwerte und Veraenderung 
GroupRate(mzTestData,TFstring="xerwstat==2&balt>=15&balt<=74",
          TFstring2="xerwstat\%in\%c(1,2)&balt>=15&balt<=74")
# Oesterreichische Bevoelkerung nach Bundesland und Geschlecht
GroupSize(mzTestData,TFstring=NULL,each="xnuts2+bsex", add_sample_size = TRUE)


\dontrun{
############################################################################################
#   Zusaetzliche Beispiele fuer DatennutzerInnen der Mikrozensus-Arbeitskraefteerhebung:   #
############################################################################################
# Quartal und zugehoeriges Vorjahrsquartal einlesen (Funktion fuer STAT-interne Nutzer)
dat <- ImportData(year=2014,quarter=4, comp_diff_lag=4)

# Oesterreichische Bevoelkerung nach Bundesland und Geschlecht
GroupSize(dat,TFstring=NULL,each="xnuts2+bsex")

# Arbeitslosenzahlen: Absolutwerte und Veraenderung
GroupSize(dat,TFstring="xerwstat==2&balt>=15&balt<=74")
# Arbeitslosenquoten: Prozentwerte und Veraenderung 
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat\%in\%c(1,2)&balt>=15&balt<=74")
# Arbeitslosenzahl vom aktuelleren der beiden Quartale
GroupSize(dat[1],TFstring="xerwstat==2&balt>=15&balt<=74")
# Arbeitslosenquote vom weniger aktuellen Quartal
GroupRate(dat[2],TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat\%in\%c(1,2)&balt>=15&balt<=74")

# Absolutwerte und Veraenderung fuer jede Auspraegung von xerwstat eingeschraenkt auf 15-74-Jaehrige
GroupSize(dat,TFstring="balt>=15&balt<=74",each="xerwstat")
# Prozentwerte (bezogen auf Gesamtbevoelkerung) und Veraenderung 
# fuer jede Auspraegung von xerwstat eingeschraenkt auf 15-74-Jaehrige
GroupRate(dat,TFstring="balt>=15&balt<=74",each="xerwstat")
# Arbeitslosenzahlen: Absolutwerte und Veraenderung fuer jedes Bundesland
GroupSize(dat,TFstring="xerwstat==2&balt>=15&balt<=74",each="xnuts2")

# Arbeitslosenquote: Prozentwerte und Veraenderung pro Bundesland
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat\%in\%c(1,2)&balt>=15&balt<=74",each="xnuts2")
# Arbeitslosenquote: Prozentwerte und Veraenderung pro Geschlecht
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat\%in\%c(1,2)&balt>=15&balt<=74",each="bsex")
# Arbeitslosenquote: Prozentwerte und Veraenderung pro Bundesland X Geschlecht
GroupSize(dat,TFstring="xerwstat==2&balt>=15&balt<=74",each="xnuts2+bsex")
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat\%in\%c(1,2)&balt>=15&balt<=74",each="xnuts2+bsex")

# Haushalte: Quartal und Vorquartal einlesen.
dat <- ImportData(year=2014,quarter=4, comp_diff_lag=1, hh=TRUE)
# Absolutwerte: Anzahl der Hauptmietwohnungen ohne gueltiger Kostenangabe.
GroupSize(dat,TFstring="wrecht==3")
GroupSize(dat,TFstring="wrecht2\%in\%c(1:3)")
}

}
\seealso{
\code{\link{ImportData},\link{IndivImportData},\link{ImportAndMerge},\link{GetLabels},\link{Total},\link{Mean},\link{export}}
}
