% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.R
\name{mz.friedman1}
\alias{mz.friedman1}
\title{generate modified friedman1 data set: 5 real, 5 uniform noise, 5 correlated with real (rs>.97), p-15 normal noise}
\usage{
mz.friedman1(n = 100, p = 55)
}
\arguments{
\item{n}{sample size}

\item{p}{total predictors (>15)}
}
\value{
returns a list of (y,x,xy) y is numeric vector, x is numeric data frame, already z scored (to matrix: data.matrix(x)), xy is a data frame with xy
}
\description{
generate modified friedman1 data set: 5 real, 5 uniform noise, 5 correlated with real, p-15 normal noise
}
\note{
y = 10sin(πx1x2) + 20(x3 − 0.5)^2 + 10x4 + 5x5 + e
}
