% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.R
\name{mz.getModelInfo}
\alias{mz.getModelInfo}
\title{getModelInfo}
\usage{
mz.getModelInfo(model = NULL, regex = TRUE, ...)
}
\arguments{
\item{model}{train method name}

\item{regex}{T/F, matching with regex or not}
}
\value{
returns an invisible data frame
}
\description{
\code{\link[caret]{getModelInfo}}
}
\seealso{
\url{http://topepo.github.io/caret/available-models.html}
}
