% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.R
\name{mz.precess}
\alias{mz.precess}
\title{preProcess + predict itself}
\usage{
mz.precess(x, ...)
}
\arguments{
\item{x}{x}
}
\description{
reProcess + predict itself
}
\note{
short for the following:
\cr prep <- preProcess(x, method = c("BoxCox","center", "scale"))
\cr x <- predict(prep, x)
}
