% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.R
\name{mz.sbfself}
\alias{mz.sbfself}
\title{sbf selected features: among resamples, univariate vars selected, how many times (or percentage) appeared in the resamples. col sorted by appearance percentage}
\usage{
mz.sbfself(sbfObj)
}
\description{
sbf selected features: among resamples, univariate vars selected, how many times (or percentage) appeared in the resamples. col sorted by appearance percentage
}
