\name{mzR-package}
\alias{mzR-package}
\alias{mzR}
\docType{package}
\title{
parser for mzXML, mzData and mzML files (mass spectrometry data)
}
\description{
  The mzR package is a library purely for accessing mass spectrometry
  data in a wide range of formats. Several backend libraries are used,
  such as the ISB random acces parser (RAMP) for mass
  spectrometry mzXML, mzData and mzML files. The package
  contains the original RAMP code written by the ISB, and a
  subset of the proteowizard library for mzML.
}
\details{


Further information is available in the following vignette:
\tabular{ll}{
\code{mzR} \tab mzR, Ramp, mzXML, mzData, mzML (source, pdf)\cr
}
}
\author{
Bernd Fischer, Steffen Neumann, Laurent Gatto

Maintainers: Bernd Fischer <bernd.fischer@embl.de>,
Steffen Neumann <sneumann@ipb-halle.de>,
Laurent Gatto <lg390@cam.ac.uk>
}

\references{
  Nat Biotechnol. 2012 Oct 10;30(10):918-20. doi: 10.1038/nbt.2377.
  A cross-platform toolkit for mass spectrometry and proteomics.
  Chambers MC, Maclean B, Burke R, Amodei D, Ruderman DL, Neumann S,
  Gatto L, Fischer B, Pratt B, Egertson J, Hoff K, Kessner D, Tasman N,
  Shulman N, Frewen B, Baker TA, Brusniak MY, Paulse C, Creasy D,
  Flashner L, Kani K, Moulding C, Seymour SL, Nuwaysir LM, Lefebvre B,
  Kuhlmann F, Roark J, Rainer P, Detlev S, Hemenway T, Huhmer A,
  Langridge J, Connolly B, Chadick T, Holly K, Eckels J, Deutsch EW,
  Moritz RL, Katz JE, Agus DB, Maccoss M, Tabb DL, Mallick P.
  http://www.ncbi.nlm.nih.gov/pubmed/23051804
}
\keyword{package, file}
