% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzR-package.R
\docType{package}
\name{mzR}
\alias{mzR-package}
\alias{mzR}
\title{mzR: Microcensus Tools}
\description{
R package to compute estimates and their errors for Austrian Microcensus (labour force survey)
data.
}
\section{Test data}{

\link{mzTestData}
}

\section{Import of data from STAT internal sources}{

Import data from the file system provided by STAT. These functions will not be operational for
external users.

\link{ImportData}, \link{ImportDataListQT}, \link{ImportAndMerge}
}

\section{Individual data import}{

Import spss and csv files based on custom paths.

\link{IndivImportData}
}

\section{Compute estimates and uncertainty}{

Create objects of the class \code{mzR} which contain estimates, confidence intervals and more based
on calibrated bootstrapping.

\link{GroupRate}, \link{GroupSize}, \link{Mean}, \link{Total}, \link{Median}
}

\section{Methods for class \code{mzR}}{

\link{print.mzR}, \link{as.data.frame.mzR}, \link{plot.mzR}, \link{AddVariable}, \link{GetLabels}
}

\section{Calculate tables}{

Calculate tables for excel export

\link{MakeTable}, \link{MakeQT}, \link{MakeAKETimeInstantsTable}
}

\section{Excel export}{

\link{export}, \link{FillExcelTemplate}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statistikat/mzR}
}

}
\author{
\strong{Maintainer}: Angelika Meraner \email{angelika.meraner@statistik.gv.at}

Authors:
\itemize{
  \item Angelika Meraner \email{angelika.meraner@statistik.gv.at}
  \item Gregor de Cillia
  \item Alexander Kowarik
}

}
\keyword{internal}
