% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.mzR}
\alias{plot.mzR}
\title{Visualisiere ein mzR Objekt}
\usage{
\method{plot}{mzR}(x, ...)
}
\arguments{
\item{x}{Ein Objekt der Klasse \code{mzR}.}

\item{...}{Ungenuetzt.}
}
\description{
Standarplots fuer \code{mzR} Objecte. Nur anwendbar, wenn das Objekt bootsrtap Replikate beinhaltet,
ansonsten wird eine Warnung geworfen. Siehe auch \link{getReplicates}.
}
\examples{
data(mzTestData)
mzObj <- Mean(mzTestData, TFstring = "xerwstat==1&balt >= 15&balt <= 74",
              var = "estund+dtstd", replicates = TRUE)
plot(mzObj)
}
