% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mzR}
\alias{print.mzR}
\title{Ergebnisse der Schaetz-und Fehlerrechnungsfunktionen werden in abgewandelter
Form in der R-Konsole ausgegeben.}
\usage{
\method{print}{mzR}(x, ...)
}
\arguments{
\item{x}{Ergebnis von GroupSize, GroupRate, Total oder Mean.}

\item{...}{currently ignored}
}
\value{
Falls \code{each} nicht auf \code{NULL} gesetzt wird, wird eine
Liste mit den Ergebnissen ausgegeben. Deren Laenge variiert je nach
Dimension der \code{each}-Variable. Ist \code{each=NULL}, so wird ein
Dataframe ausgegeben.
}
\description{
Funktion liest Ergebnisse von \code{GroupSize} oder \code{GroupRate} ein und
gibt sie in der R-Konsole in etwas abgewandelter Form aus bzw. als Objekt
zurueck.
}
\details{
...
}
