% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.R
\name{varImpRFE}
\alias{varImpRFE}
\title{only for RFE: pass rfeObj, plot optVar averaged importance across OptSized resamples, return an invisible df}
\usage{
varImpRFE(rfeObj, plot = T)
}
\description{
only for RFE: pass rfeObj, plot optVar averaged importance across OptSized resamples, return an invisible df
}
