% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{getReplicates}
\alias{getReplicates}
\title{Extrahiere die Bootstrap Replikla aus einem mzR Objekt}
\usage{
getReplicates(x)
}
\arguments{
\item{x}{Ein Objekt der Klasse \code{mzR}. Typischerweise generiert durch \code{GroupRate}, \code{Groupsize},
\code{Median}, \code{Mean} oder \code{Total}.}
}
\value{
Tabelle mit Bootstrapreplika. Die Spalten entsprechen den Gruppierungsvariablen der
Auswertung, falls vorhanden (siehe das Argument \code{each} in \code{GroupRate}, \code{Median}, etc.). Die Zeilen
enstsprechen den Bootstrapgewichten (typischerweise 500 St체ck).
}
\description{
Falls ein \code{mzR} Objekt mit dem Parameter \code{replicate = TRUE} erzeugt wurde, lassen sich mit dieser
Funktion alle Sch채tzwerte zur체ckgeben.
}
\examples{
library(dplyr)  ## F체r \%>\%

######################## Beispiel 1: Durschnittliche Arbeitsstunden #############################

dat <- ImportData(year = 2014, quarter = 4)
mzObj <- Mean(dat, TFstring = "xerwstat==1&balt >= 15&balt <= 74", 
              var = "estund*13+dtstd*13", replicates = TRUE)
replicates <- getReplicates(mzObj)$replicates

hist(replicates, main = "Durschnittlich geleistete Arbeitsstunden, 95\% KI", freq = FALSE,
     col = "lightblue", xlab = "Durchschnittliche Wochenstunden in Replika")
abline(v = quantile(replicates, c(.025, 0.975)), col = "red", lwd = 2)
lines(density(replicates), col = "darkblue", lwd = 2)

########################### Beispiel 2: Wohnkosten nach Geschlecht ##############################

mzObj2 <- Mean(dat, TFstring = "xerwstat==1&balt >= 15&balt <= 74", var = "wkges", 
               replicates = TRUE, each = "bsex")
getReplicates(mzObj2) \%>\% tidyr::gather(Geschlecht, wk) \%>\% 
  mutate(Geschlecht = recode(Geschlecht, bsex_1 = "M", bsex_2 = "W")) \%>\%
  ggplot(aes(Geschlecht, wk, fill = Geschlecht)) + 
  geom_boxplot(outlier.shape = NA, alpha = 0.5) +
  geom_jitter(aes(col = Geschlecht), alpha = 0.5) +
  ylab("Durchschnittliche Wohnkosten in Replika")
  
###################### Beispiel 3: Arbeitslosenquote nach Bundesland ############################

mzObj3 <- GroupSize(dat,TFstring="xerwstat==2&balt>=15&balt<=74", replicates = TRUE, 
                    each = "xnuts2")
getReplicates(mzObj3) \%>\% tidyr::gather(Bundesland, unemployment) \%>\% 
ggplot(aes(Bundesland, unemployment, fill = Bundesland)) + 
  geom_boxplot(outlier.shape = NA, alpha = 0.5) +
  geom_jitter(aes(col = Bundesland), alpha = 0.05) +
  ylab("Arbeislosenzahlen in Bootstrap-Replika")
}
