% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDatalistQT.R
\name{ImportDataListQT}
\alias{ImportDataListQT}
\title{Sehr spezifische Funktion die MZ-Daten einliest um damit in weiterer Folge Tabellen im
MZ-Quartalsbericht-Format zu erstellen (hausintern).}
\usage{
ImportDataListQT(timeInstant, nbw = NULL, whichVar = NULL,
  weightDecimals = 2, ImportAndMerge = TRUE, curr_inFile = NULL,
  prev_inFile = NULL, mergeBy = c("asbper", "ajahr", "amonat"))
}
\arguments{
\item{timeInstant}{numerischer Vektor mit 2 Elementen: c(jahr, quartal).
Hier gibt man den Zeitpunkt an auf den sich alle Ergebnisse im weitesten
Sinn beziehen sollen, also i.d.R. das aktuellste Quartal.}

\item{nbw}{numerischer Wert: Anzahl an Bootstrap-Gewichten die eingelesen
werden soll (z.B. um Rechenzeit beim Aufsetzen der Tabellen zu verkuerzen).}

\item{whichVar}{Character (vector) oder NULL. Falls ungleich NULL, Character Vektor mit Variable(n) aus
dem dg7-Mikrozensus-File die im Output-File enthalten sein sollen. Die
uebrigen Variablen werden weggelassen. Default ist NULL, dabei werden alle
Variablen behalten.}

\item{weightDecimals}{Numerischer Wert oder NULL. Anzahl der Nachkommastellen der Stichprobengewichte, 
gerundet nach SPSS RND Logik (0.5 bwz. -0.5 wird dabei immer "weg von 0" gerundet). 
Falls NULL, werden die Gewichte nicht gerundet.}

\item{ImportAndMerge}{TRUE/FALSE ob die Funktion \link{ImportAndMerge} angewendet werden soll. 
Bei der Defaulteinstellung \code{ImportAndMerge=TRUE} und \code{curr_inFile=NULL} wird \emph{Daten_ab2004_QuartPub.sav}
zu den MZ-Daten gemerged.}

\item{curr_inFile}{Pfad der Datei die eingelesen und zu den MZ-Daten gemerged werden soll (bezogen auf
den aktuelleren der beiden Zeitpunkte falls prev_inFile ungleich NULL).
Eingelesen werden koennen Files vom Typ .sav, .csv und .csv.gz.}

\item{prev_inFile}{Falls ungleich NULL, Pfad der Datei die eingelesen und zu den MZ-Daten gemerged werden
soll (bezogen auf den weniger aktuellen Zeitpunkt). Eingelesen werden koennen
Files vom Typ .sav, .csv und .csv.gz.}

\item{mergeBy}{Character Vektor mit Variable(n) nach denen gemerged werden
soll (default=c("asbper","ajahr","amonat")).}
}
\value{
Output ist eine Liste deren Elemente jeweils MZ-Daten enthalten 
die die selbe Grundstruktur haben wie der Output aus der Fimltopm \link{ImportData}.
}
\description{
Funktion liest MZ-Daten (dg7) automatisch aus dem STAT-Filemanagement ein und fuehrt diese Daten
mit dem derzeit fuer die Quartalsberichtsproduktion verwendeten SPSS-File zusammen (siehe Details).
}
\details{
Anzugeben ist bei dieser Funktion der Referenzzeitpunkt \code{timeInstant} des MZ-Quartalsberichts.
Die MZ-Daten (dg7) werden (derzeit) defaultmaessig (\code{ImportAndMerge=TRUE})  
ueber \code{mergeBy = c("asbper","ajahr","amonat")} mit \emph{Daten_ab2004_QuartPub.sav} 
gemerged. Dieses File wird quartalsweise vom Fachbereich erzeugt und immer im selben Ordner abgelegt.
Sollte man mit einem anderen File mergen wollen, so kann man die entsprechenden File-Pfade in 
\code{curr_inFile} und \code{prev_inFile} spezifizieren aber auch \code{mergeBy} anpassen, analog zu \link{ImportAndMerge}.
Des Weiteren koennen die einzulesenden MZ-Daten mit den Funktionsparametern
\code{nbw}, \code{whichVar} und \code{weightDecimals} angepasst werden, siehe \link{ImportData}. 
Dadurch kann man z.B. vermeiden, dass Variablen die sowohl in den MZ-Daten als auch in 
\emph{Daten_ab2004_QuartPub.sav} vorkommen doppelt eingelesen werden.
}
\examples{
\dontrun{
## Lesen Daten fuer den AKE-QT-Referenzzeitpunkt 2014q4 ein.
# Fuer Testzwecke (um Rechenzeit zu sparen) schraenken wir die 
# Anzahl der Bootstrapgewichte ein auf 5.
# Ausserdem wollen wir aus den Original-MZ-Daten (dg7) nur die 
# Variable \\emph{rbpkin} behalten.#' 

datalist <- ImportDataListQT(timeInstant=c(2014,4), nbw=5, whichVar=c("rbpkin"))

}

}
\seealso{
\code{\link{MakeQT},\link{MakeTable},\link{FillExcelTemplate},\link{ImportData},\link{ImportAndMerge}}
}

