\name{mzR-package}
\alias{mzR-package}
\alias{mzR}
\docType{package}
\title{
parser for mzXML, mzData and mzML files (mass spectrometry data)
}
\description{
  The mzR package is a library purely for accessing mass spectrometry
  data in a wide range of formats. Several backend libraries are used,
  such as the ISB random acces parser (RAMP) for mass
  spectrometry mzXML, mzData and mzML files. The package
  contains the original RAMP code written by the ISB, and a
  subset of the proteowizard library for mzML.
}
\details{


Further information is available in the following vignette:
\tabular{ll}{
\code{mzR} \tab mzR, Ramp, mzXML, mzData, mzML (source, pdf)\cr
}
}
\author{
Bernd Fischer, Steffen Neumann, Laurent Gatto

Maintainers: Bernd Fischer <bernd.fischer@embl.de>,
Steffen Neumann <sneumann@ipb-halle.de>,
Laurent Gatto <lg390@cam.ac.uk>
}
\references{
  http://dx.doi.org/10.1074/mcp.R110.000133
}
\keyword{package, file}
