% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzR-package.R
\docType{package}
\name{mzR}
\alias{mzR}
\alias{mzR-package}
\title{mzR: Microcensus tools}
\description{
R package to compute estimates and their errors for Austrian Microcensus (labour force survey) data.
}
\section{Testdaten}{

\link{mzTestData}
}

\section{Import von STAT internen Daten}{

\link{ImportData}, \link{ImportDataListQT}
}

\section{Individueller Datenimport}{

\link{IndivImportData}
}

\section{Schaetz- und Fehlerrechnungsfunktionen}{

\link{GroupRate}, \link{GroupSize}, \link{Mean}, \link{Total}
}

\section{Excel export}{

\link{MakeTable}, \link{MakeQT}, \link{FillExcelTemplate}
}

\section{Utility Funktionen}{

\link{AddVariable}, \link{GetLabels}
}

