% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.mzR}
\alias{plot.mzR}
\title{Visualisiere ein mzR Objekt}
\usage{
\method{plot}{mzR}(x)
}
\arguments{
\item{x}{Ein Objekt der Klasse \code{mzR}.}
}
\description{
Standarplots für \code{mzR} Objecte. Nur anwendbar, wenn das Objekt bootsrtap Replikate beinhaltet,
ansonsten wird eine Warnung geworfen. Siehe auch \link{getReplicates}.
}
\examples{
dat <- ImportData(year = 2014, quarter = 4)
mzObj <- Mean(dat, TFstring = "xerwstat==1&balt >= 15&balt <= 74", 
              var = "estund*13+dtstd*13", replicates = TRUE)
plot(mzObj)
}
