% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalMean.R
\name{Median}
\alias{Median}
\alias{Mean}
\alias{Total}
\alias{Mean}
\alias{Median}
\title{Schaetz- und Fehlerrechnungsfunktionen fuer numerische Variablen.}
\usage{
Median(x, TFstring = NULL, each = NULL, var, negativeZero = TRUE,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)

Mean(x, TFstring = NULL, each = NULL, var, negativeZero = TRUE,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)

Total(x, TFstring = NULL, each = NULL, var, negativeZero = TRUE,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)
}
\arguments{
\item{x}{MZ Daten - Output von Funktion \link{ImportData}.}

\item{TFstring}{Character oder NULL: Logische Einschraenkung der Gruppe. Falls NULL, gilt keine Einschraenkung.}

\item{each}{Character oder NULL: Name der Variable nach der getrennt berechnet werden soll.
Hier koennen auch mehrere Variablen angegeben werden und zwar in einem character string, getrennt durch ein +, siehe Examples.}

\item{var}{character: Name der numerischen Variable oder Formel zur Berechnung dieser
z.b. 'x+y'.}

\item{negativeZero}{Logical: Wenn TRUE, werden die negativen Werte der numerischen
Variable auf 0 gesetzt.}

\item{thousands_separator}{Logical: Wenn TRUE, werden Tausendertrennzeichen
angezeigt.}

\item{digits}{Numerischer Wert: Anzahl der Nachkommastellen im angezeigten Ergebnis. Default
ist 2.}

\item{replicates}{Fürge einen Vektor aus Schätzwerten zum Output hinzu? Die Anzahl der Schätzwerte
pro Gruppe in \code{each} entspricht der Anzahl der Bootstrapreplikate (typischerweise 500).
Siehe auch \link{getReplicates}.}
}
\value{
Output ist ein Objekt der Klasse \code{mzR}.
}
\description{
Funktion berechnet Totalwerte oder Mittelwerte einer numerischen Variable
und die zugehoerigen Fehler.
}
\details{
Wiedergegeben wird der Schaetzer \code{est}, der Stichprobenfehler
\code{sd}, der Variationskoeffizient \code{cv} und die untere/obere Grenze
des 95\% Konfidenzintervalls \code{cil_2.5\%}/\code{ciu_97.5\%}.
Die Fehler werden mit Hilfe von Replikationsgewichten aus einem Bootstrapverfahren
berechnet, d.h. \code{sd} entspricht der Standardabweichung der mit den Bootstrapgewichten
berechneten Schaetzwerte, \code{cil_2.5\%} und \code{ciu_97.5\%} sind die
entsprechenden 2.5\% und 97.5\% Quantile und \code{cv=sd/est}.
}
\section{Functions}{
\itemize{
\item \code{Median}: Median. Robuste alternative zu \code{Mean}

\item \code{Mean}: Arithmetisches Mittel
}}

\examples{
# Daten laden (oder Daten einlesen mit ImportData() bzw. IndivImportData())
data(mzTestData)

# Geleistete Arbeitsstunden: Absolutwerte
Total(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund*13+dtstd*13")
# Geleistete Arbeitsstunden: Absolutwerte in Millionen
Total(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="(estund+dtstd)*13/10^6")

#Durchschnittlich geleistete Arbeitsstunden
Mean(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund+dtstd")
#Durchschnittlich geleistete Arbeitsstunden nach Geschlecht
Mean(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund+dtstd", each="bsex")
#Durchschnittlich geleistete Arbeitsstunden nach Bundesland und Geschlecht
Mean(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund+dtstd",each="xnuts2+bsex")

}
\seealso{
\code{\link{ImportData},\link{IndivImportData},\link{ImportAndMerge},
\link{GetLabels},\link{GroupSize},\link{GroupRate},\link{export.mzR}}
}
